fat_gui/requirements.txt
необходимые библиотеки, устанавливать через pip install
requests
Pillow
tkinter
reportlab